bitrate

+++ x2mp4.pl
@@ -11,11 +11,18 @@ if(!$out){
 
 my $info=`ffmpeg -i "$in" 2>&1 | grep fps`;
 my ($rate) = $info=~/(\d+) kb/; 
+if(!$rate){
+    $info=`ffmpeg -i "$in" 2>&1 | grep bitrate`;
+    ($rate) = $info=~/(\d+) kb/; 
+}
 
-print "video: $in -> $out, rate : $rate kb/s\n";
-my $cmd =qq[ffmpeg -i "$in" -y -strict -2 -b:${rate}k -o "$out"];
+if($rate){
+    print "video: $in -> $out, rate : $rate kb/s\n";
+    my $cmd =qq[ffmpeg -i "$in" -y -strict -2 -b:${rate}k -o "$out"];
 
-eval { 
-    `$cmd >/dev/null 2>&1`; 
-    unlink($in) if(-f $out and -s $out);
-};
+    eval { 
+        `$cmd >/dev/null 2>&1`; 
+        unlink($in) if(-f $out and -s $out);
+    };
+}
+
